fix #1, improve SRAM restore performance, fix minor work indicator bug

+++ src/main.c
@@ -26,7 +26,7 @@
 #include "ws/hardware.h"
 #include "ws/system.h"
 
-volatile uint8_t vbl_ticks;
+volatile uint16_t vbl_ticks;
 
 __attribute__((interrupt))
 void vblank_int_handler(void) {
